requirements.txt
streamlit
pymupdf
pdf2image
openai
Pillow
reportlab
python-dotenv
np
